+++ src/renderer/src/services/messageStreaming/callbacks/citationCallbacks.ts
@@ -1,0 +1,109 @@
+import type { ExternalToolResult } from '@renderer/types'
+import { CitationMessageBlock, MessageBlockStatus, MessageBlockType } from '@renderer/types/newMessage'
+import { createCitationBlock } from '@renderer/utils/messageUtils/create'
+import { findMainTextBlocks } from '@renderer/utils/messageUtils/find'
+
+import { BlockManager } from '../BlockManager'
+
+interface CitationCallbacksDependencies {
+  blockManager: BlockManager
+  assistantMsgId: string
+  getState: any
+}
+
+export const createCitationCallbacks = (deps: CitationCallbacksDependencies) => {
+  const { blockManager, assistantMsgId, getState } = deps
+
+  // 内部维护的状态
+  let citationBlockId: string | null = null
+
+  return {
+    onExternalToolInProgress: async () => {
+      const citationBlock = createCitationBlock(assistantMsgId, {}, { status: MessageBlockStatus.PROCESSING })
+      citationBlockId = citationBlock.id
+      await blockManager.handleBlockTransition(citationBlock, MessageBlockType.CITATION)
+    },
+
+    onExternalToolComplete: (externalToolResult: ExternalToolResult) => {
+      if (citationBlockId) {
+        const changes: Partial<CitationMessageBlock> = {
+          response: externalToolResult.webSearch,
+          knowledge: externalToolResult.knowledge,
+          status: MessageBlockStatus.SUCCESS
+        }
+        blockManager.smartBlockUpdate(citationBlockId, changes, MessageBlockType.CITATION, true)
+      } else {
+        console.error('[onExternalToolComplete] citationBlockId is null. Cannot update.')
+      }
+    },
+
+    onLLMWebSearchInProgress: async () => {
+      if (blockManager.hasInitialPlaceholder) {
+        blockManager.lastBlockType = MessageBlockType.CITATION
+        citationBlockId = blockManager.initialPlaceholderBlockId!
+        const changes = {
+          type: MessageBlockType.CITATION,
+          status: MessageBlockStatus.PROCESSING
+        }
+        blockManager.smartBlockUpdate(citationBlockId, changes, MessageBlockType.CITATION)
+      } else {
+        const citationBlock = createCitationBlock(assistantMsgId, {}, { status: MessageBlockStatus.PROCESSING })
+        citationBlockId = citationBlock.id
+        await blockManager.handleBlockTransition(citationBlock, MessageBlockType.CITATION)
+      }
+    },
+
+    onLLMWebSearchComplete: async (llmWebSearchResult: any) => {
+      const blockId = citationBlockId || blockManager.initialPlaceholderBlockId
+      if (blockId) {
+        const changes: Partial<CitationMessageBlock> = {
+          type: MessageBlockType.CITATION,
+          response: llmWebSearchResult,
+          status: MessageBlockStatus.SUCCESS
+        }
+        blockManager.smartBlockUpdate(blockId, changes, MessageBlockType.CITATION)
+
+        const state = getState()
+        const existingMainTextBlocks = findMainTextBlocks(state.messages.entities[assistantMsgId])
+        if (existingMainTextBlocks.length > 0) {
+          const existingMainTextBlock = existingMainTextBlocks[0]
+          const currentRefs = existingMainTextBlock.citationReferences || []
+          const mainTextChanges = {
+            citationReferences: [...currentRefs, { blockId, citationBlockSource: llmWebSearchResult.source }]
+          }
+          blockManager.smartBlockUpdate(existingMainTextBlock.id, mainTextChanges, MessageBlockType.MAIN_TEXT, true)
+        }
+
+        if (blockManager.hasInitialPlaceholder) {
+          citationBlockId = blockManager.initialPlaceholderBlockId
+        }
+      } else {
+        const citationBlock = createCitationBlock(
+          assistantMsgId,
+          {
+            response: llmWebSearchResult
+          },
+          {
+            status: MessageBlockStatus.SUCCESS
+          }
+        )
+        citationBlockId = citationBlock.id
+
+        const state = getState()
+        const existingMainTextBlocks = findMainTextBlocks(state.messages.entities[assistantMsgId])
+        if (existingMainTextBlocks.length > 0) {
+          const existingMainTextBlock = existingMainTextBlocks[0]
+          const currentRefs = existingMainTextBlock.citationReferences || []
+          const mainTextChanges = {
+            citationReferences: [...currentRefs, { citationBlockId, citationBlockSource: llmWebSearchResult.source }]
+          }
+          blockManager.smartBlockUpdate(existingMainTextBlock.id, mainTextChanges, MessageBlockType.MAIN_TEXT, true)
+        }
+        await blockManager.handleBlockTransition(citationBlock, MessageBlockType.CITATION)
+      }
+    },
+
+    // 暴露给外部的方法，用于textCallbacks中获取citationBlockId
+    getCitationBlockId: () => citationBlockId
+  }
+}
