+++ src/renderer/src/services/messageStreaming/callbacks/textCallbacks.ts
@@ -1,0 +1,75 @@
+import { WebSearchSource } from '@renderer/types'
+import { CitationMessageBlock, MessageBlock, MessageBlockStatus, MessageBlockType } from '@renderer/types/newMessage'
+import { createMainTextBlock } from '@renderer/utils/messageUtils/create'
+
+import { BlockManager } from '../BlockManager'
+
+interface TextCallbacksDependencies {
+  blockManager: BlockManager
+  getState: any
+  assistantMsgId: string
+  getCitationBlockId: () => string | null
+}
+
+export const createTextCallbacks = (deps: TextCallbacksDependencies) => {
+  const { blockManager, getState, assistantMsgId, getCitationBlockId } = deps
+
+  // 内部维护的状态
+  let accumulatedContent = ''
+  let mainTextBlockId: string | null = null
+
+  return {
+    onTextStart: async () => {
+      if (blockManager.hasInitialPlaceholder) {
+        blockManager.lastBlockType = MessageBlockType.MAIN_TEXT
+        const changes = {
+          type: MessageBlockType.MAIN_TEXT,
+          content: accumulatedContent,
+          status: MessageBlockStatus.STREAMING
+        }
+        mainTextBlockId = blockManager.initialPlaceholderBlockId!
+        blockManager.smartBlockUpdate(mainTextBlockId, changes, MessageBlockType.MAIN_TEXT, true)
+      } else if (!mainTextBlockId) {
+        const newBlock = createMainTextBlock(assistantMsgId, accumulatedContent, {
+          status: MessageBlockStatus.STREAMING
+        })
+        mainTextBlockId = newBlock.id
+        await blockManager.handleBlockTransition(newBlock, MessageBlockType.MAIN_TEXT)
+      }
+    },
+
+    onTextChunk: async (text: string) => {
+      const citationBlockId = getCitationBlockId()
+      const citationBlockSource = citationBlockId
+        ? (getState().messageBlocks.entities[citationBlockId] as CitationMessageBlock).response?.source
+        : WebSearchSource.WEBSEARCH
+
+      accumulatedContent += text
+
+      if (mainTextBlockId) {
+        const blockChanges: Partial<MessageBlock> = {
+          content: accumulatedContent,
+          status: MessageBlockStatus.STREAMING,
+          citationReferences: citationBlockId ? [{ citationBlockId, citationBlockSource }] : []
+        }
+        blockManager.smartBlockUpdate(mainTextBlockId, blockChanges, MessageBlockType.MAIN_TEXT)
+      }
+    },
+
+    onTextComplete: async (finalText: string) => {
+      if (mainTextBlockId) {
+        const changes = {
+          content: finalText,
+          status: MessageBlockStatus.SUCCESS
+        }
+        blockManager.smartBlockUpdate(mainTextBlockId, changes, MessageBlockType.MAIN_TEXT, true)
+        mainTextBlockId = null
+        accumulatedContent = ''
+      } else {
+        console.warn(
+          `[onTextComplete] Received text.complete but last block was not MAIN_TEXT (was ${blockManager.lastBlockType}) or lastBlockId is null.`
+        )
+      }
+    }
+  }
+}
