+++ src/renderer/src/services/messageStreaming/callbacks/thinkingCallbacks.ts
@@ -1,0 +1,71 @@
+import { MessageBlock, MessageBlockStatus, MessageBlockType } from '@renderer/types/newMessage'
+import { createThinkingBlock } from '@renderer/utils/messageUtils/create'
+
+import { BlockManager } from '../BlockManager'
+
+interface ThinkingCallbacksDependencies {
+  blockManager: BlockManager
+  assistantMsgId: string
+}
+
+export const createThinkingCallbacks = (deps: ThinkingCallbacksDependencies) => {
+  const { blockManager, assistantMsgId } = deps
+
+  // 内部维护的状态
+  let accumulatedThinking = ''
+  let thinkingBlockId: string | null = null
+
+  return {
+    onThinkingStart: async () => {
+      if (blockManager.hasInitialPlaceholder) {
+        blockManager.lastBlockType = MessageBlockType.THINKING
+        const changes = {
+          type: MessageBlockType.THINKING,
+          content: accumulatedThinking,
+          status: MessageBlockStatus.STREAMING,
+          thinking_millsec: 0
+        }
+        thinkingBlockId = blockManager.initialPlaceholderBlockId!
+        blockManager.smartBlockUpdate(thinkingBlockId, changes, MessageBlockType.THINKING, true)
+      } else if (!thinkingBlockId) {
+        const newBlock = createThinkingBlock(assistantMsgId, accumulatedThinking, {
+          status: MessageBlockStatus.STREAMING,
+          thinking_millsec: 0
+        })
+        thinkingBlockId = newBlock.id
+        await blockManager.handleBlockTransition(newBlock, MessageBlockType.THINKING)
+      }
+    },
+
+    onThinkingChunk: async (text: string, thinking_millsec?: number) => {
+      accumulatedThinking += text
+
+      if (thinkingBlockId) {
+        const blockChanges: Partial<MessageBlock> = {
+          content: accumulatedThinking,
+          status: MessageBlockStatus.STREAMING,
+          thinking_millsec: thinking_millsec || 0
+        }
+        blockManager.smartBlockUpdate(thinkingBlockId, blockChanges, MessageBlockType.THINKING)
+      }
+    },
+
+    onThinkingComplete: (finalText: string, final_thinking_millsec?: number) => {
+      if (thinkingBlockId) {
+        const changes = {
+          type: MessageBlockType.THINKING,
+          content: finalText,
+          status: MessageBlockStatus.SUCCESS,
+          thinking_millsec: final_thinking_millsec || 0
+        }
+        blockManager.smartBlockUpdate(thinkingBlockId, changes, MessageBlockType.THINKING, true)
+        thinkingBlockId = null
+        accumulatedThinking = ''
+      } else {
+        console.warn(
+          `[onThinkingComplete] Received thinking.complete but last block was not THINKING (was ${blockManager.lastBlockType}) or lastBlockId is null.`
+        )
+      }
+    }
+  }
+}
